year 3 here we go again

+++ PATema2/src/p3.cpp
@@ -8,33 +8,33 @@ using namespace std;
 
 #define INF 1000000000
 
-int n,m,t;
-vector<tuple<int,int,int>> a[300];
-int p[120][120], cost[300];
-int u,v,c,type;
+long long n,m,t;
+vector<tuple<long long,long long,long long>> a[300];
+long long p[120][120], cost[300];
+long long u,v,c,type;
 
 int main(){
 	ifstream in("p3.in");
 	ofstream out("p3.out");
 	in >> n >> m >> t;
-	for(int i = 0; i < m; i++){
+	for(long long i = 0; i < m; i++){
 		in >> u >> v >> c >> type;
 		a[u].push_back(make_tuple(v,c,type));
 		a[v].push_back(make_tuple(u,c,type));		
 	}
-	for(int i = 1; i <= t; i++)
-		for(int j = 1; j<= t; j++)
+	for(long long i = 1; i <= t; i++)
+		for(long long j = 1; j<= t; j++)
 			in >> p[i][j];
-	for(int i = 0 ; i <= n; i++) cost[i] = INF;
-	set<tuple<int,int,int>> heap;
+	for(long long i = 0 ; i <= n; i++) cost[i] = INF;
+	set<tuple<long long,long long,long long>> heap;
 	cost[n] = 0;
 	heap.insert(make_tuple(cost[n], n, -1));
 	while(heap.size()){
-		int costFather, father, typeFather;
+		long long costFather, father, typeFather;
 		tie(costFather, father, typeFather) = *(heap.begin());
 		heap.erase(heap.begin());
-		for(int i = 0; i < a[father].size(); i++){
-			int son, costEdge, typeSon;
+		for(long long i = 0; i < a[father].size(); i++){
+			long long son, costEdge, typeSon;
 			tie(son, costEdge, typeSon) = a[father][i];
 			if(cost[son] > costFather + costEdge + (typeFather != -1 ? p[typeFather][typeSon] : 0)){
 				cost[son] = costFather + costEdge + (typeFather != -1 ? p[typeFather][typeSon] : 0);
